requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
decord>=0.6.0
einops>=0.6.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
tensorboard>=2.13.0
pyyaml>=6.0
tqdm>=4.65.0
opencv-python>=4.7.0
